Split location utils as separate crate
