+++ oak_functions/examples/weather_lookup/location_utils/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "location_utils"
+version = "0.1.0"
+authors = ["Conrad Grobler <grobler@google.com>"]
+edition = "2018"
+license = "Apache-2.0"
