+++ oak_functions/examples/weather_lookup/location_utils/src/tests.rs
@@ -1,0 +1,23 @@
+//
+// Copyright 2021 The Project Oak Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use super::*;
+
+#[test]
+fn test_index_key_round_trip() {
+    let initial = IndexKey { row: 30, col: -10 };
+    let result = IndexKey::from_bytes(&initial.to_bytes());
+    assert_eq!(initial, result);
+}
